test(runtime): isolate provider TLS regression probes
